BIG PHYSICS REVAMP

+++ include/memory.h
@@ -32,7 +32,7 @@ extern u8 curr_level_height;
 extern u16 *level_pointer[2];
 
 extern u64 scroll_x;
-extern u32 scroll_y;
+extern u64 scroll_y;
 
 extern s8 scroll_y_dir;
