+++ include/player.h
@@ -20,7 +20,7 @@ enum SPEEDS {
 };
 
 extern u64 player_x;
-extern u32 player_y;
+extern u64 player_y;
 
 extern u16 relative_player_x;
 extern u16 relative_player_y;
@@ -28,8 +28,8 @@ extern u16 relative_player_y;
 extern u16 player_width;
 extern u16 player_height;
 
-extern s16 player_x_speed;
-extern s16 player_y_speed;
+extern s32 player_x_speed;
+extern s32 player_y_speed;
 
 extern s16 gravity;
 extern u8 gravity_dir;
