+++ src/memory.c
@@ -28,7 +28,7 @@ u16 *level_pointer[LEVEL_LAYERS];
 // Scroll variables
 // in subpixels
 u64 scroll_x = 0;
-u32 scroll_y = 0;
+u64 scroll_y = 0;
 
 // 0 : up | 1 : down
 s8 scroll_y_dir = 0;
